Made presence of Numpy during compile optional.
