+++ tools/python/src/dlib.cpp
@@ -20,6 +20,7 @@ void bind_shape_predictors();
 void bind_correlation_tracker();
 void bind_face_recognition();
 void bind_cnn_face_detection();
+void bind_numpy_returns();
 
 #ifndef DLIB_NO_GUI_SUPPORT
 void bind_gui();
@@ -53,6 +54,7 @@ BOOST_PYTHON_MODULE(dlib)
     bind_correlation_tracker();
     bind_face_recognition();
     bind_cnn_face_detection();
+    bind_numpy_returns();
 #ifndef DLIB_NO_GUI_SUPPORT
     bind_gui();
 #endif
